now can check for correct password to provide session token.  need to work with frontend team to implement and verify that this really works
